Added attributed text support
Items are now initialized with either a String answer, or an NSAttributedText answer. Attributed takes precedence if both are present on the same item. An array of items may contain a mix of both. Note: The text color of attributed text must be set on the text itself. Added a property: answerTintColor which controls the text color of detected links.